Correct type warning
Resolves this error:

In file included from src/main.cpp:18:
lib/BQ27427/src/BQ27427.h:34:28: warning: unsigned conversion from 'int' to 'uint8_t' {aka 'unsigned char'} changes value from '12848' to '48' [-Woverflow]
   34 |           uint8_t chemId = BQ27427_CHEM_ID_A); // Lithium Cobalt Oxide (LiCoO₂)
      |                            ^~~~~~~~~~~~~~~~~
lib/BQ27427/src/BQ27427.h:34:28: warning: unsigned conversion from 'int' to 'uint8_t' {aka 'unsigned char'} changes value from '12848' to '48' [-Woverflow]
Compiling .pio/build/esp32-c3-devkitm-1/lib6b3/BQ27427/BQ27427.cpp.o
In file included from lib/BQ27427/src/BQ27427.cpp:14:
lib/BQ27427/src/BQ27427.h:34:28: warning: unsigned conversion from 'int' to 'uint8_t' {aka 'unsigned char'} changes value from '12848' to '48' [-Woverflow]
   34 |           uint8_t chemId = BQ27427_CHEM_ID_A); // Lithium Cobalt Oxide (LiCoO₂)

+++ src/BQ27427.cpp
@@ -17,7 +17,7 @@ https://github.com/sparkfun/SparkFun_BQ274xx_Arduino_Library
  ************************** Initialization Functions *************************
  *****************************************************************************/
 // Initializes class variables
-BQ27427::BQ27427(uint16_t capacity, uint8_t deviceAddress, uint16_t designVoltage, uint16_t taperCurrent, uint16_t terminateVoltage, uint8_t chemId)
+BQ27427::BQ27427(uint16_t capacity, uint8_t deviceAddress, uint16_t designVoltage, uint16_t taperCurrent, uint16_t terminateVoltage, uint16_t chemId)
 {
 	_wire.addr = deviceAddress;
 	_wire.bus = &_bq27427;
